// src/types/history/HistoryAction.tsx
export type HistoryAction =
    | "CREATED_TRANSACTION"
    | "CREATED_FAMILY"
    | "DELETE_FAMILY"
    | "UPDATED_FAMILY"
    | "CREATED_PURCHASE"
    | "UPDATED_PURCHASE"
    | "CLOSE_PURCHASE"
    | "DELETED_PURCHASE"
    | "CREATED_PRODUCT"
    | "UPDATED_PRODUCT"
    | "DELETED_PRODUCT"
    | "ADDED_MEMBER"
    | "REMOVED_MEMBER"
    | "CHANGE_MEMBER"
    | "EXIT_MEMBER"
    | "CREATED_CATEGORY"
    | "UPDATED_CATEGORY"
    | "DELETED_CATEGORY";